Use external precompiled built of SFML
Instead of using fetchcontent to get the SFML sources and link with
the targets SML provides, rely on SML already being installed on the
system with a package manager or by downloading the prebuilt binaries.

This is the more standard way to do it, so also built the Thor library
as a shared library which doesn't require any changes to the existing
sources.

That seems to be much closer to how SFML and Thor are intended to be
used.

Also ensure the shared libraries are in the build directory when
running on Windows and MacOs so this is hopefully easy to run the
executable after being built.