Make player class inherit from sf::Sprite
Since the player class is basically a sprite, make it inherit from the
sprite class which seems like a common thing to do in SFML anyway.

This simplifies things quite a bit and means we can get rid of the
silly custom bounds() function and just use getGlobalBounds() instead
which is much closer to how SFML is intended to be used (I guess).