Add a status panel
Show the name of the player, the remaining lives of the player and the
name of the current level at the bottom of the play area. Similar to
the original Pang game.

This requires adding a font, so use the nice Atarian System font from
here:

https://www.1001fonts.com/sf-atarian-system-font.html

Which should be completely free to use for this purpose and looks
quite nice as well.

For now names and live counts are hardcoded, but they're not really
used for anything anyway right now.